+++ src/lib/user.ts
@@ -1,0 +1,6 @@
+import { request } from '.'
+
+export const getUser = async () => {
+  const res = await request.get('/users')
+  return res?.data?.user
+}
